+++ webapp/src/app/resource/components/simple-section/section.definitions.ts
@@ -1,28 +1,28 @@
-import { SectionOptions } from './section.component';
+import { SimpleSectionOptions } from './simple-section.component';
 import { DocumentSectionType } from '../outline/document-outline.model';
 
-export const commentsSectionOptions: SectionOptions = {
+export const commentsSectionOptions: SimpleSectionOptions = {
   title: 'Things To Consider',
   fontAwesomeIcon: 'fa-comment-dots',
   iconBackground: 'blue',
   sectionType: DocumentSectionType.ThingsToConsider
 };
 
-export const instructionalUseOptions: SectionOptions = {
+export const instructionalUseOptions: SimpleSectionOptions = {
   title: 'Instructional Use',
   sbdlIcon: 'steps',
   iconBackground: 'yellow',
   sectionType: DocumentSectionType.Subsection
 };
 
-export const strategyInActionOptions: SectionOptions = {
+export const strategyInActionOptions: SimpleSectionOptions = {
   title: 'Strategy In Action',
   fontAwesomeIcon: 'fa-universal-access',
   iconBackground: 'green',
   sectionType: DocumentSectionType.Subsection
 };
 
-export const stepByStepOptions: SectionOptions = {
+export const stepByStepOptions: SimpleSectionOptions = {
   title: 'Step-by-Step',
   sbdlIcon: 'steps',
   iconBackground: 'blue',
