+++ webapp/src/app/resource/components/simple-section/simple-section.component.html
