+++ webapp/src/app/resource/professional/content/professional-content.component.html
@@ -30,11 +30,11 @@
   (sectionLoaded)="addDocumentSection($event)">
 </sbdl-attachments>
 
-<sbdl-section
+<sbdl-simple-section
   [contentHtml]="resource.thingsToConsider"
-  [options]="commentsOptions"
+  [options]="commentsSectionOptions"
   (sectionLoaded)="addDocumentSection($event)">
-</sbdl-section>
+</sbdl-simple-section>
 
 <sbdl-formative-list
   [strategies]="resource.formativeAssessmentStrategies"
