+++ webapp/src/app/resource/resource-content.component.ts
@@ -1,6 +1,6 @@
 import { EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { Map } from 'immutable';
-import { commentsSectionOptions } from './components/section/section.definitions';
+import { commentsSectionOptions } from './components/simple-section/section.definitions';
 import { Resource } from '../data/resource/model/resource.model';
 import { Note } from '../data/notes/model/note.model';
 import { DocumentOutline, DocumentSection, DocumentSectionType } from './components/outline/document-outline.model';
@@ -11,7 +11,7 @@ import { DocumentOutline, DocumentSection, DocumentSectionType } from './compone
  */
 export class ResourceContentComponent implements OnInit {
 
-    commentsOptions = commentsSectionOptions;
+    commentsSectionOptions = commentsSectionOptions;
 
     @Input()
     resource: Resource;
