+++ webapp/src/app/resource/resource.module.ts
@@ -1,8 +1,14 @@
-import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { SbdlCommonModule } from '../common/common.module';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+
 import { PipesModule } from '../pipes/pipes.module';
 import { NotesModule } from '../notes/notes.module';
+import { SbdlCommonModule } from '../common/common.module';
+
+import { ResourceHostDirective } from './resource-host.directive';
+import { ResourceTypeStrategyComponent } from './resource-type-strategy.component';
+
 import { ActionsComponent } from './components/actions/actions.component';
 import { AttachmentsComponent } from './components/attachments/attachments.component';
 import { DifferentiationComponent } from './instructional/differentiation/differentiation.component';
@@ -22,15 +28,11 @@ import { OutlineComponent } from './components/outline/outline.component';
 import { EnhancedPrintingComponent } from './components/outline/enhanced-printing/enhanced-printing.component';
 import { PrintCheckboxComponent } from './components/outline/enhanced-printing/print-checkbox.component';
 import { ResourceComponent } from './resource.component';
-import { ResourceHostDirective } from './resource-host.directive';
 import { ResourcePropertiesComponent } from './components/resource-properties/resource-properties.component';
-import { ResourceTypeStrategyComponent } from './resource-type-strategy.component';
-import { SectionComponent } from './components/section/section.component';
+import { SimpleSectionComponent } from './components/simple-section/simple-section.component';
 import { StepByStepComponent } from './components/step-by-step/step-by-step.component';
 import { StepComponent } from './components/step-by-step/step.component';
 import { StrategyReferenceListComponent } from './components/strategy-reference-list/strategy-reference-list.component';
-import { AppRoutingModule } from '../app-routing.module';
-import { RouterModule } from '@angular/router';
 
 export function getResourceComponents(resourceComponent: ResourceComponent) {
   return resourceComponent;
@@ -60,17 +62,18 @@ export function getResourceComponents(resourceComponent: ResourceComponent) {
     ResourceHostDirective,
     ResourcePropertiesComponent,
     ResourceTypeStrategyComponent,
-    SectionComponent,
+    SimpleSectionComponent,
     StepByStepComponent,
     StepComponent,
     StrategyReferenceListComponent
   ],
   imports: [
     CommonModule,
+    RouterModule,
+
     PipesModule,
     NotesModule,
-    SbdlCommonModule,
-    RouterModule
+    SbdlCommonModule
   ],
   entryComponents: [
     // Resource components are loaded dynamically so they need to be explicitly set here
